Move DecimalFormats into StringUtils.
